% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{change_door}
\alias{change_door}
\title{Change door strategy.}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{The}{function utilizes opened.door and a.pick and sets the stay value
to T.}
}
\value{
The The function returns a numeric value representing 1 door,
neither the door in select_door or opened_goat_door.
}
\description{
\code{change_door()} simulates a strategy of switching from the initial pick
to the remaining door.
}
\details{
This function simulates the change the door strategy. In this strategy
the contestant decides to switch from their initial pick chosen
in the select_door function. The final pick is neither the door selected
in select_door or open_goat_door. The if statement identifies this as
final.pick and is returned if the change door strategy is chosen.
}
\examples{
  change_door()
}
