% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{determine_winner}
\alias{determine_winner}
\title{Determine winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{This}{function is dependent on final.pick and game.}
}
\value{
The function returns a length 3 character vector
indicating the positions of goats and the car.
}
\description{
\code{determine_winner()}determines if the contestant chose the door
with the car or goat.
}
\details{
This function identifies if the contestant won or lost the game.
If the final pick was the door associated with the car the
contestant wins. If the final pick was door associated with the
goat the contestant loses.
}
\examples{
  determine_winner()
}
