% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{open_goat_door}
\alias{open_goat_door}
\title{Opens a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{This}{function uses game and a.pick as parameters.}
}
\value{
The function returns a numeric value representing 1 door, the
numeric value associated with a door with a goat.
}
\description{
\code{open_goat_door()} simulates the reveal of one of the doors with
a goat.
}
\details{
Following the picking of the initial door two doors remain. One
of these doors is a door with a goat behind it. In this round
one of the doors with a goat is identified. This function takes into
account the door picked in the select_door function and identifies
a goat door. If the initial pick was a door with a goat the remaining
goat door is opened. If the initial pick had the car one of the goat
doors is chosen at random.
}
\examples{
  return(opened.door)
}
