% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{play_game}
\alias{play_game}
\title{Play Monty Hall Game}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns character vectors of the stay and switch strategies
indicating win or lose.
}
\description{
\code{play_game()} plays the Monty Hall game.
}
\details{
This function plays the Monty Hall game. The function creates the game,
chooses the initial door, opens a door with a goat, and runs the
simulation for both staying with the initial door and switching doors.
The function assesses the outcomes of both switch and stay strategies.
The game provides results for either switching or staying with the
initial door.
}
\examples{
  play_game()
}
