% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{play_n_games}
\alias{play_n_games}
\title{Play n Number of Game}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{Setting}{the n of the function.}
}
\value{
Table of numeric values indicating the percentage of times a win occurred.
}
\description{
\code{play_n_games()} plays the Monty Hall game as many times as the
simulator wants to play.
}
\details{
The function allows the simulator to choose an n for the number of times
the simulator wishes to simulate the Monty Hall game. Choosing the n
allows the simulator to choose a number and calculate the ideal
winning strategy. Running the simulation a large number of times
will indicate whether the strategy of stay or switch is most
likely to result in a win.
}
\examples{
play_n_games()
}
