% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-problem2.R
\name{select_door}
\alias{select_door}
\title{Select the first door.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a numeric value representing 1 door, the
door picked by the contestant.
}
\description{
\code{select_door()} generates the first door our of three that
the contestant picks.
}
\details{
The initial door picked is the first move in the game.
In this step The contestant picks the initial door.
The door is either door number 1, door number 2, or door number 3.
This pick sets up the rest of the game where a goat door is shown and
the contestant can choose to either stay with the door chosen in this
step or switch doors.
}
\examples{
  return(a.pick)
}
